refactor: move savannah logic to service
