+++ src/app/mini-games/savannah/pages/savannah-mini-game/savannah-mini-game.component.html
@@ -15,7 +15,7 @@
   </div>
 
   <ng-container *ngSwitchCase="gameState.SETTING">
-    <app-difficultyy-input (startWithDifficulty)="getWords($event)"></app-difficultyy-input>
+    <app-difficultyy-input (startWithDifficulty)="startGameWithDifficulty($event)"></app-difficultyy-input>
   </ng-container>
 
   <ng-container *ngSwitchCase="gameState.PREP">
